+++ stdcommands/leaveserver/leaveserver.go
@@ -16,7 +16,7 @@ var Command = &commands.YAGCommand{
 	HideFromHelp:         true,
 	RequiredArgs:         1,
 	Arguments: []*dcmd.ArgDef{
-		{Name: "server", Type: dcmd.Int},
+		{Name: "server", Type: dcmd.BigInt},
 	},
 	RunFunc: util.RequireOwner(func(data *dcmd.Data) (interface{}, error) {
 		err := common.BotSession.GuildLeave(data.Args[0].Int64())
