+++ stdcommands/topservers/topservers.go
@@ -19,7 +19,7 @@ var Command = &commands.YAGCommand{
 		{Name: "Skip", Help: "Entries to skip", Type: dcmd.Int, Default: 0},
 	},
 	ArgSwitches: []*dcmd.ArgDef{
-		{Name: "id", Type: dcmd.Int},
+		{Name: "id", Type: dcmd.BigInt},
 	},
 	RunFunc: func(data *dcmd.Data) (interface{}, error) {
 		skip := data.Args[0].Int()
